Create selection_sort.c
this c program is for selection sort.
In selection sort, all the steps are same as insertion sort. i.e We divide the array in two parts sorted and unsorted.
Only difference is, in selection sort at first the sorted array is empty. Then we need to select the minimum element out of the unsorted array in push that into the empty array.
Time complexity is O(n^2) for both best and worst case.